+++ SimpleSheets/SimpleSheets.Data/Models/WorkHistory.cs
@@ -7,6 +7,6 @@ namespace SimpleSheets.Data.Models
     public class WorkHistory
     {
         public string TImeSheetEntryDate{ get; set; }
-        public int NoOfHours { get; set; }
+        public float NoOfHours { get; set; }
     }
 }
